fix: Fix integer type on PagedResponse, add getters and setters for TransferPeriod

+++ sctp-api/src/main/java/org/cga/sctp/api/core/pagination/PagedResponse.java
@@ -54,7 +54,7 @@ public class PagedResponse<T> {
         this.items = page.toList();
     }
 
-    public PagedResponse(int page, int totalItems, int totalPages, List<T> items) {
+    public PagedResponse(int page, long totalItems, int totalPages, List<T> items) {
         this.page = page;
         this.totalItems = totalItems;
         this.totalPages = totalPages;
